requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-knox==4.1.0
djangorestframework==3.12.1
Pillow==8.0.1
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
